Move API coupling to relevant modules
