--- app/modules/api/feed/actionTypes.js
@@ -1,5 +1,0 @@
-// @flow
-
-export const FETCH_FEED: 'feed/FETCH_FEED' = 'feed/FETCH_FEED';
-export const FETCH_FEED_SUCCESS: 'feed/FETCH_FEED_SUCCESS' = 'feed/FETCH_FEED_SUCCESS';
-export const FETCH_FEED_FAILURE: 'feed/FETCH_FEED_FAILURE' = 'feed/FETCH_FEED_FAILURE';
