--- app/modules/api/index.js
@@ -1,11 +1,0 @@
-// @flow
-
-import feedSagas from './feed/sagas';
-
-const sagas = [
-  ...feedSagas,
-];
-
-export default {
-  sagas,
-};
