+++ app/modules/feed-items/actionTypes.js
@@ -7,6 +7,10 @@ export const ADD: 'feedItems/ADD' = 'feedItems/ADD';
 
 export const ADD_ERROR: 'feedItems/ADD_ERROR' = 'feedItems/ADD_ERROR';
 
+export const FETCH_FEED: 'feed/FETCH_FEED' = 'feed/FETCH_FEED';
+export const FETCH_FEED_SUCCESS: 'feed/FETCH_FEED_SUCCESS' = 'feed/FETCH_FEED_SUCCESS';
+export const FETCH_FEED_FAILURE: 'feed/FETCH_FEED_FAILURE' = 'feed/FETCH_FEED_FAILURE';
+
 export type AddAction = {
   type: typeof ADD,
   payload: {
