+++ app/modules/feed-items/constants.js
@@ -1,4 +1,5 @@
 // @flow
 
-// API endpoint
+export const GRAVATAR_SIZE = 64;
+
 export const ENDPOINT = '/notifications';
