+++ app/modules/feed-items/sagaHelpers.js
@@ -3,7 +3,7 @@
 import { call, put } from 'redux-saga/effects';
 
 import { FETCH_FEED_SUCCESS, FETCH_FEED_FAILURE } from './actionTypes';
-import * as Api from './api';
+import Api from './api';
 
 export function* fetch() {
   try {
